Removed warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@10866 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: ab0b046853acbbba88fb68a0a9345e393f26a65c

+++ costmodel/de.uniol.inf.is.odysseus.costmodel.operator.datasrc/src/de/uniol/inf/is/odysseus/costmodel/operator/datasrc/impl/histogram/FreedmanDiaconisRule.java
@@ -9,8 +9,8 @@ public class FreedmanDiaconisRule implements IIntervalCountEstimator {
 		int dataSize = data.size();
 
 		// determine number of bins...
-		int lowerQuartilePos = (int) ((double) dataSize * 0.25);
-		int upperQuartilePos = (int) ((double) dataSize * 0.75);
+		int lowerQuartilePos = (int) (dataSize * 0.25);
+		int upperQuartilePos = (int) (dataSize * 0.75);
 		double min = data.get(0);
 		double max = data.get(dataSize - 1);
